+++ src/irmin-unix/resolver.ml
@@ -22,12 +22,21 @@ let global_option_section = "COMMON OPTIONS"
 
 module Conf = Irmin.Backend.Conf
 
+let string_or_option ty v =
+  match Irmin.Type.of_string ty v with
+  | Error e -> (
+      let x = Format.sprintf "{\"some\": %s}" v in
+      match Irmin.Type.of_string ty x with
+      | Error _ ->
+          let y = Format.sprintf "{\"some\": \"%s\"}" v in
+          Irmin.Type.of_string ty y |> Result.map_error (fun _ -> e)
+      | v -> v)
+  | v -> v
+
 let pconv t =
   let pp = Irmin.Type.pp t in
   let parse s =
-    match Irmin.Type.of_string t s with
-    | Ok x -> `Ok x
-    | Error (`Msg e) -> `Error e
+    match string_or_option t s with Ok x -> `Ok x | Error (`Msg e) -> `Error e
   in
   (parse, pp)
 
@@ -386,7 +395,6 @@ let config_term =
     Arg.info ~docv:"OPTIONS" ~docs:global_option_section ~doc
       [ "opt"; "options" ]
   in
-
   let root_key = Conf.root spec in
   let config_path_key =
     Conf.key ~spec ~docs:global_option_section ~docv:"PATH"
@@ -465,13 +473,7 @@ let load_config_file_with_defaults root config_path (store, hash, contents) =
         match Conf.Spec.find_key spec k with
         | Some (Irmin.Backend.Conf.K k) ->
             let v = json_of_yaml v |> Yojson.Basic.to_string in
-            let v =
-              match Irmin.Type.of_json_string (Conf.ty k) v with
-              | Error _ ->
-                  let v = Format.sprintf "{\"some\": %s}" v in
-                  Irmin.Type.of_json_string (Conf.ty k) v |> Result.get_ok
-              | Ok v -> v
-            in
+            let v = string_or_option (Conf.ty k) v |> Result.get_ok in
             Conf.add config k v
         | None -> (
             match k with
@@ -506,17 +508,7 @@ let from_config_file_with_defaults root config_path (store, hash, contents) opts
                 | None -> invalid_arg ("opt: " ^ k)
             in
             let ty = Conf.ty key in
-            let v =
-              match Irmin.Type.of_string ty v with
-              | Error _ -> (
-                  let x = Format.sprintf "{\"some\": %s}" v in
-                  match Irmin.Type.of_string (Conf.ty key) x with
-                  | Error _ ->
-                      let y = Format.sprintf "{\"some\": \"%s\"}" v in
-                      Irmin.Type.of_string (Conf.ty key) y |> Result.get_ok
-                  | Ok v -> v)
-              | Ok v -> v
-            in
+            let v = string_or_option ty v |> Result.get_ok in
             let config = Conf.add config key v in
             config)
           config (List.flatten opts)